+++ .gitignore
@@ -1,0 +1,8 @@
+# Environments
+.env
+env
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
